+++ CHANGES.md
@@ -1,6 +1,9 @@
 CHANGES
 =======
 
+- `AuthorizationFailReason`
+    * Added `AuthorizationFailReason_INVALID_TARGET`.
+
 v1.0.5 (2019-09-25)
 -------------------