+++ CHANGES.md
@@ -11,6 +11,9 @@ CHANGES
     * Added `Resources` field.
     * Added `Purpose` field.
 
+- `BackchannelAuthenticationCompleteRequest` struct
+    * Added `IdtHeaderParams` field.
+
 v1.0.5 (2019-09-25)
 -------------------