+++ CHANGES.md
@@ -49,6 +49,10 @@ CHANGES
     * Added `Htm` field.
     * Added `Htu` field.
 
+- `IntrospectionResponse` struct
+    * Added `Resources` field.
+    * Added `AccessTokenResources` field.
+
 v1.0.5 (2019-09-25)
 -------------------